+++ CHANGELOG.md
@@ -26,10 +26,13 @@
   - volume support for API300
   - volume_template support for API300
   - enclosure support for API300. Also added `:patch` action
+  - Added support to SAS Logical Interconnects for API300::Synergy
   - logical_switch support for API300
   - interconnect support for API300
   - server_hardware support for API300. Also added `:patch` action
   - switch support for API300 C7000. Added `:patch` action
+  - server_profile support for API300
+  - server_profile_template support for API300
 
 ### 1.1.0
   - Add support for client ENV variables